% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rcmip-inputs.R
\name{generate_rcmip_inputs}
\alias{generate_rcmip_inputs}
\title{Generate RCMIP input files}
\usage{
generate_rcmip_inputs(targetfile = NULL)
}
\value{
\code{targetfile}, invisibly
}
\description{
Generate RCMIP input files
}
