% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/halocarbon-info.R
\name{halocarbon_data}
\alias{halocarbon_data}
\title{Complete halocarbon dataframe}
\usage{
halocarbon_data()
}
\value{
\code{data.frame} of halocarbon info
}
\description{
Complete halocarbon dataframe
}
\author{
Alexey Shiklomanov
}
