% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/halocarbon-info.R
\name{halocarbon_info}
\alias{halocarbon_info}
\title{Information about a specific halocarbon}
\usage{
halocarbon_info(halocarbon)
}
\value{
\code{data.frame} with a single row containing halocarbon info
}
\description{
Information about a specific halocarbon
}
\author{
Alexey Shiklomanov
}
