% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/implied-emissions.R
\name{implied_emissions}
\alias{implied_emissions}
\title{Calculate implied emissions from concentration data}
\usage{
implied_emissions(concdata, component)
}
\arguments{
\item{concdata}{(data.frame) Input concentration data}

\item{component}{(character) Name of component}
}
\value{

}
\description{
Calculate implied emissions from concentration data
}
\author{
Alexey Shiklomanov
}
