% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rcmip-core.R
\name{rcmip_ini}
\alias{rcmip_ini}
\title{RCMIP default INI file path}
\usage{
rcmip_ini()
}
\value{
Full path to RCMIP default INI file
}
\description{
RCMIP default INI file path
}
\author{
Alexey Shiklomanov
}
