% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rcmip-outputs.R
\name{rcmip_outputs}
\alias{rcmip_outputs}
\title{Retrieve results with RCMIP variable names and units}
\usage{
rcmip_outputs(core, ...)
}
\arguments{
\item{core}{Hector core (as returned by \code{\link[=run_scenario]{run_scenario()}})}

\item{...}{Additional arguments to \code{\link[=fetchvars2]{fetchvars2()}}}
}
\value{

}
\description{
Retrieve results with RCMIP variable names and units
}
\author{
Alexey Shiklomanov
}
