% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/run-probability.R
\name{run_probability}
\alias{run_probability}
\title{Run scenario with parameter uncertainty}
\usage{
run_probability(scenario, n = 1000, ...)
}
\arguments{
\item{scenario}{Name of scenario}

\item{n}{Number of ensembles to run}

\item{...}{Additional arguments to \code{\link[=set_variable]{set_variable()}}}
}
\value{
Nested \code{data.frame} of outputs
}
\description{
Run scenario with parameter uncertainty
}
\author{
Alexey Shiklomanov
}
