% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/run-scenario.R
\name{run_scenario}
\alias{run_scenario}
\title{Run an RCMIP Scenario using Hector}
\usage{
run_scenario(scenario)
}
\arguments{
\item{scenario}{Name of scenario}
}
\description{
Run an RCMIP Scenario using Hector
}
\author{
Alexey Shiklomanov
}
