% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/run-probability.R
\name{run_with_param}
\alias{run_with_param}
\title{Run Hector with specific parameter values}
\usage{
run_with_param(core, pS, pdiff, palpha, dates = 1750:2100,
  include_params = TRUE, .pb = NULL, ...)
}
\arguments{
\item{core}{Hector \code{core} object}

\item{pS}{Equilibrium climate sensitivity (\code{S})}

\item{pdiff}{Heat diffusivity (\code{diff})}

\item{palpha}{Aerosol scaling factor (\code{alpha})}

\item{include_params}{Logical. If \code{TRUE} (default), include parameter values
in output object}

\item{.pb}{Optional \code{\link[progress:progress_bar]{progress::progress_bar()}} bar object. If \code{NULL}, ignore.}

\item{...}{Additional arguments to \code{\link[=fetchvars2]{fetchvars2()}}}
}
\value{
\code{data.frame} of results
}
\description{
Run Hector with specific parameter values
}
\author{
Alexey Shiklomanov
}
