% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/run-scenario.R
\name{set_scenario}
\alias{set_scenario}
\title{Add scenario data to a Hector core}
\usage{
set_scenario(hc, scenario, ...)
}
\arguments{
\item{hc}{Hector core (see \code{\link[hector:newcore]{hector::newcore()}})}

\item{scenario}{Name of scenario}

\item{...}{Additional arguments to \code{\link[=set_variable]{set_variable()}}}
}
\value{
Hector core (invisibly)
}
\description{
Add scenario data to a Hector core
}
\author{
Alexey Shiklomanov
}
