% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rplot.R
\name{rplot}
\alias{rplot}
\title{Quick plot of Hector results}
\usage{
rplot(results)
}
\arguments{
\item{results}{Hector results, as produced by \code{\link[hector:fetchvars]{hector::fetchvars()}}, or a
Hector core, in which case default variables will be read in.}
}
\value{
\code{ggplot2} plot object
}
\description{
Quick plot of Hector results
}
\author{
Alexey Shiklomanov
}
