% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/run-scenario.R
\name{run_scenario}
\alias{run_scenario}
\title{Run an RCMIP Scenario using Hector}
\usage{
run_scenario(scenario, control = FALSE)
}
\arguments{
\item{scenario}{Name of scenario}

\item{control}{(Logical) If \code{TRUE}, assume a control run and set more stuff
to zero.}
}
\value{
Hector core object at the run end date
}
\description{
Run an RCMIP Scenario using Hector
}
\author{
Alexey Shiklomanov
}
