% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/run-scenario.R
\name{run_scenario}
\alias{run_scenario}
\title{Run an RCMIP Scenario using Hector}
\usage{
run_scenario(scenario, cmip6_model = NULL, outfile = NULL, ...)
}
\arguments{
\item{scenario}{Name of scenario}

\item{cmip6_model}{CMIP6 model parameter set to use. If \code{NULL} (default), use
Hector defaults.}

\item{outfile}{Output file. Default is \verb{output/zz-raw-output/<scenario>-<model>.csv}}

\item{...}{Additional arguments to \code{\link[=set_variable]{set_variable()}}}
}
\value{
Hector core object at the run end date
}
\description{
Run an RCMIP Scenario using Hector
}
\author{
Alexey Shiklomanov
}
