% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/set-variable.R
\name{set_variable}
\alias{set_variable}
\title{Set Hector variable to RCMIP data}
\usage{
set_variable(
  core,
  input_data,
  varname = NULL,
  hector_vars = rcmip2hector_df(),
  interpolate = TRUE
)
}
\arguments{
\item{core}{Hector core object}

\item{input_data}{\code{data.frame} of RCMIP inputs for a specific scenario.}

\item{varname}{RCMIP variable name. Defaults to unique \code{Variable} in
\code{input_data}.}

\item{hector_vars}{RCMIP to Hector variable conversion table}

\item{interpolate}{(Logical) If \code{TRUE} (default), interpolate incomplete time
series using \code{\link[stats:approxfun]{stats::approxfun()}}}
}
\value{
\code{core}, invisibly
}
\description{
Set Hector variable to RCMIP data
}
\author{
Alexey Shiklomanov
}
